# tests/baseline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOAR Lab Management API",
    "description": "API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Login endpoint - returns JWT token.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "summary": "Verify Auth",
        "description": "Verify JWT token endpoint.",
        "operationId": "verify_auth_auth_verify_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyAuthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/backup/create": {
      "post": {
        "summary": "Create Backup",
        "description": "Create a backup.",
        "operationId": "create_backup_backup_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBackupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/backup/restore": {
      "post": {
        "summary": "Restore Backup",
        "description": "Restore a backup.",
        "operationId": "restore_backup_backup_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreBackupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tests/run": {
      "post": {
        "summary": "Run Tests",
        "description": "Run tests.",
        "operationId": "run_tests_tests_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupRequest": {
        "properties": {
          "backup_name": {
            "type": "string",
            "title": "Backup Name",
            "description": "Name of the backup file to restore"
          }
        },
        "type": "object",
        "required": [
          "backup_name"
        ],
        "title": "BackupRequest",
        "description": "Backup request model.\n\nSpecifies the backup file name for create or restore operations."
      },
...
      "CreateBackupResponse": {
        "properties": {
          "backup_name": {
            "type": "string",
            "title": "Backup Name",
            "description": "Created backup name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Operation status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "type": "object",
        "required": [
          "backup_name",
          "message"
        ],
        "title": "CreateBackupResponse",
        "description": "Create backup response model.\n\nContains the result of a backup creation operation."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Login request model.\n\nContains username and password fields for authentication."
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Authentication token"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type",
            "default": "Bearer"
          }
        },
        "type": "object",
        "required": [
          "token",
          "message"
        ],
        "title": "LoginResponse",
        "description": "Login response model.\n\nContains the JWT token and related metadata."
      },
...
      "RestoreBackupResponse": {
        "properties": {
          "backup_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backup Name",
            "description": "Restored backup name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Operation status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RestoreBackupResponse",
        "description": "Restore backup response model.\n\nContains the result of a backup restore operation."
      },
      "RunRequest": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Test category (e.g., 'unit', 'integration', 'all')"
          }
        },
        "type": "object",
        "required": [
          "category"
        ],
        "title": "RunRequest",
        "description": "Test execution request model.\n\nSpecifies the test category to run."
      },
      "RunResults": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Test category"
          },
          "passed": {
            "type": "integer",
            "title": "Passed",
            "description": "Number of passed tests"
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "description": "Number of failed tests"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "description": "Number of skipped tests"
          },
          "coverage": {
            "type": "number",
            "title": "Coverage",
            "description": "Test coverage percentage"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "description": "Test output"
          },
          "duration": {
            "type": "number",
            "title": "Duration",
            "description": "Test execution duration in seconds"
          }
        },
        "type": "object",
        "required": [
          "category",
          "passed",
          "failed",
          "skipped",
          "coverage",
          "output",
          "duration"
        ],
        "title": "RunResults",
        "description": "Test results model.\n\nContains pass/fail/skip counts and coverage info."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyAuthResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "Whether the token is valid"
          },
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User",
            "description": "User information"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "user"
        ],
        "title": "VerifyAuthResponse",
        "description": "Verify authentication response model.\n\nIndicates whether a token is valid and includes user info."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}